fix checksec regression under certain circumstances and version
This resolves that a failing checksec command may output something when
falling back to the next version that will be tried. Additionally this
introduces a last option to try to call checksec before failing as some
distributions (like Arch Linix) package checksec under /usr/bin/checksec
and expect --file as command line option.
The returned output is decoded to ensure converting the bytes() object
to a proper string representation needed for python3, under python2 this
should simply be a no-op.
The line ending of the print command is stripped as the check_output
already returns a newline when not failing.