fixed potential wrong config param type resolution (#102)
This fixes an issue that can potentially happen with unordered dicts
in python3. A boolean typed config parameter can be handles as
integer dependin on the order of iteration in getParam(value).

The reason for this is because True/False are both, bool and int
at the same time, however its not the other way around.