Fix Ehdr and Phdr endian switching
Also fixes regression - issue #156.
Before this changes we were creating ElfXX_Ehdr and ElfXX_Phdr structures once during initialization using `pwndbg.ctypes.Structure` as base class (which was `ctypes.LittleEndianStructure` - as the default `pwndbg.arch.endian` is 'little').
We were thinking that the proper Ehdr/Phdr is set based on architecture (`pwndbg.arch.ptrsize`) and endian (as we updated the `pwndbg.ctypes.Structure` attribute to either `ctypes.LittleEndianStructure` or `ctypes.BigEndianStructure`).
However, changing the `pwndbg.ctypes.Structure` didn't really refresh the ELfXX_Ehdr and ElfXX_Phdr structures, so they still inherited from the old `pwndbg.ctypes.Structure` (which was `ctypes.LittleEndianStructure`).

Because of above, when we loaded a binary which was big endian, we failed on parsing its Ehdr and/or Phdr structures (as we just thought we updated Ehdr/Phdr to proper endian instead of really doing it; and so we interpreted Ehdr/Phdr fields as little endian).

This commit fixes the issue by removing the `pwndbg.ctypes` totally and by creating Ehdr and Phdr classes for each architecture (`pwndbg.arch.ptrsize` - 4 or 8) and endian (`pwndbg.arch.endian` - 'big' or 'little').
To make the code not so much bloated when it is possible the classes are generated.