Add pwndbg.storage and four caching API file_cache_{,with_signature_}for_{text,pickle}
Consider an imaginary file pwndbg/foo.py

@pwndbg.storage.file_cache_for_text('{0}.txt')
def ft(x):
    print(x)
    return x

The first call of ft('a') will save a file ~/.cache/pwndbg/foo/ft/a.txt .
The basename is generated from '{0}.txt'.format(x)
A later call of ft('a') will load the text from the cache.

The template can also be specified as a callable
@pwndbg.storage.file_cache_for_text(lambda x: x.replace('/', '')+'.txt')

@pwndbg.storage.file_cache_with_signature_for_text('{0}')
def fts(x):
    print(x)
    return x

The caller should provide an extra keyword argument (`signature`) to `fts`.
The first call of fts('a', signature='sig') will save a file ~/.cache/pwndbg/foo/fts/a .
A call of fts('a', signature='sig') will load the cache.
However, a call of fts('a', signature='different_sig') will recalculate as the signature is different.