Fixes #428 - pwndbg.memory.write encoding error
It seems that pwndbg.memory.write fix for Py2 introduced in 433bf231
wasn't tested properly on Py3.

In Py2 by default the `bytes` is just `str` and so doesn't accept the encoding argument.

Because of that a `from builtins import bytes` has been added.

Some more info on `builtins` module can be found here: http://python-future.org/imports.html#imports-of-builtins