Workaround for GDB bug described in #632
GDB's `up` and `down` commands trigger internal notification about
changed frame. It does not happen for `gdb.Frame.select()` which we use
in our own overrides for `up` and `down` commands.

Because of that, the `list` GDB command does not show proper source code
lines.

This can be worked around by firing `frame` command and this is what
this workaround/PR adds.

This bug has also been reported to GDB bugzilla at
https://sourceware.org/bugzilla/show_bug.cgi?id=24534