Fix remote debugging via pwntools
When we debug something remotely we add symbols from the binary path.
It turns out that GDB does it by itself, so it is not really needed.
And it seems that when we do it by ourselves, the symbols sometimes end
up being wrong (e.g. we see a dispaly of `func1+X` instead of `func2`).