command: support alternatives including sub command wrapper like pwn (#701)
* command: support alternatives including sub command wrapper like pwn

Additionally speed up the checksec logic by removing the --version
check. Simply try to shell out and use the first working option variant
as we can't easily detect the different available alternatives in a
bulletproof way.

This implementation allows to use the sub command wrapper 'pwn' to call
checksec in case pwntools has been installed using --only-use-pwn-command

To unconditionally unify the usage, the functions cmd attribute stores
the base command to execute in form of a list that can be used to
concatenate an array of options against it and pass the final list to
call_cmd

* checksec: allow a third option variant by only passing the file target

Lets try a third available version of checksec as the lowest order
variant in our cascade that works by only passing the target file
without additional specifier.