Fix inaccuracies in vis_heap_chunks() (#708)
* Fix inaccuracies in vis_heap_chunks()

Correct boundary calculation of chunks adjacent to a top chunk and
use a list of boundary addresses to do so, rather than a dictionary of
every quadword in each printed chunk. Ensure that at least 1 chunk is
printed even when a non-sensical value of 0 is passed as the count argument.

* Avoid out-of-bounds read in vis_heap_chunks()

While processing a corrupt heap or using the --naive option, the
vis_heap_chunks command could read out-of-bounds. Add a check to
avoid this scenario.

* Use chunk_flags instead of constants

To find out whether a chunk's prev_inuse bit is set, use chunk_flags() rather than anding with constants.ptmalloc.PREV_INUSE.

Co-Authored-By: Paweł Płatek <e2.8a.95@gmail.com>

* Use main_heap.size_sz

Use main_heap.size_sz in place of pwndbg.arch.ptrsize.

Co-Authored-By: Paweł Płatek <e2.8a.95@gmail.com>

* Replace a magic number with ptr_size

Co-Authored-By: Paweł Płatek <e2.8a.95@gmail.com>

* Calculate color_func variable in 1 line

Co-Authored-By: Paweł Płatek <e2.8a.95@gmail.com>

* Fix crash in get_heap_boundaries()

Don't assume that because an address was supplied to
get_heap_boundaries() that it resides on a legitimate heap. Sometimes a
user may want to print fake chunks elsewhere in memory.

Co-authored-by: Paweł Płatek <e2.8a.95@gmail.com>