Fixes #749 - stop showing pc marker in disasm loops (#750)
This commit fixes the issue described in #749.

During disasm output, we enhance the display to show additional information of the instructions.
When a future instruction executes a branch instruction (jmp/call), we fetch the next instruction based on the jmp/call target, as long as we can calculate it statically.

If we can calculate it statically, we will then display the target of the jmp/call as the next instruction, as e.g. in here:
```
 > 0x5555555545fe <main+4>    jmp    main+4 <0x5555555545fe>
    v
 > 0x5555555545fe <main+4>    jmp    main+4 <0x5555555545fe>
```

The issue is, that we mark both instructions as "current", highlighting both of them, making it a bit unambigous "where we are".

While this view is _kinda valid_ as the PC is really the same, we want to mark/hightlight only the first instruction we are on, as it is the one that is being executed right now and the program might go some other path in the future.

This commit fixes this display by simply making it so that the `nearpc` function/command used to display disasm shows the marker only once, for the first time it shows the current PC instruction.