Fixes #777 - missing pyelftools program header name
When pyelftools missed a PT_* name, invoking a `xinfo` command on a library with the missing program header crashed the `xinfo` command due to us assuming the ptype will be a string.

Instead, since pyelftoools didn't have the name, this ph is then an int.

So this commit takes this into acccount and omits those program headers that we weren't able to name due to lack of info in pyelftools.

This solution is not ideal as we might miss something, but given that we care about given program header name, I don't think it will be a big deal for us.