Allow return offsets and use it for 'start' method.
The 8f33ec4 made `pwndbg.symbol.address` to discard addresses
of symbols not mapped.

Unfortunately this broke pwndbg's `start`.

GDB's `start` puts a temporal break in `main` and pwndbg's `start` does
the same but when GDB returns the address of `main`, it returns an
offset the first time because the symbol was not mapped yet.

The offset is then discarded and pwndbg doesn't put the breakpoint when
it should.

This PR fixes pwndbg's `start` allowing `pwndbg.symbol.address` to
return offsets instead of addresses: GDB will resolve the correct
address when it builds the breakpoint and pwndbg's `start` will behave
like GDB's `start`.