Enhance Dockerfile to use the latest GDB and install a non-root user
This Dockerfile is also based on Ubuntu 20.04, but installs the very latest
version of GDB, so that things can be tested against it.

It also sets up the locale properly so that there are no locale warnings
when launching GDB.

Finally, a pwndbg user is created so that we can test things as non-root.

The main makefile targets are "all" (which builds to Docker image) and
"run" which runs the resulting Docker image.