feature(ghidra): use configurable code prefix marker for line indicator
As the source code and the decompiled sources are essentially the same
thing, lets just reuse the existing code prefix marker to indicate the
current line instead of using a hardcoded plain string.

A comment compatible marker is used before the syntax highlighter to
avoid any highlight and parsing confusion which is later replaced by the
colorized variant of the prefix marker before returning the results.

Furthermore we only replace the amount of indented spaces that is
required to fill the space for the code prefix marker.