fix(ghidra): handle PIE base address when opening the r2pipe
If we are trying to decompile a running binary which is a PIE, we need
to make sure to pass the appropriate base address to radare2 to be used
when loading a new binary.

Furthermore set io.cache to fix relocations in disassembly and avoid a
warning from the r2pipe.