Accept args given in the format string of printf-like functions
See #939

It doesn't really adjust the value format (as I thought it would when I
provide the argument type, but it turns out it's not interpreted
anywhere), so e.g. for floats you won't get anything nice as 1.5 but
still some unreadable 0x..., but at least it's a start of reading
`printf` arguments.