Improve vmmap on coredump files (#1149)
* Improve vmmap on coredump files

With this commit we now recognize coredumps better and also finally have
a simple test for vmmap commands on:
- a running binary
- on a loaded coredump file with loaded binary
- on a loaded coredump file without a loaded binary

We also stop saving vmmaps for `maintenance info sections` sections
which have a start address of 0x0. While there could potentially be a
coredump file from a binary with start=0x0, this should work in most
cases.

We could in theory do a slighty better: we could take the vmmap at 0 and
try to read memory from it. However, I am not sure if it is a good idea
to try such memory read?

* remove unused import

* add missing crash_simple.asm

* fix vmmap coredump test on different ubuntu mem layouts

* use /proc/$pid/maps for vmmap tests

* fix formatting

* fix import

* fix test

* fix test

* fix test

* fix lint

* fix test

* fix test

* fix test

* fix test

* fix lint

* another fixup for ubuntu 22.04

* another fixup for ubuntu 22.04

* lint